// src/shared/api/generated/models/supportInspectorEventsListParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type SupportInspectorEventsListParams = {
  from: string;
  to: string;
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  /**
   * @maxLength 4096
   */
  cursor?: string;
};