// src/shared/api/generated/models/supportLeadActivityParams.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type SupportLeadActivityParams = {
  /**
   * @maxLength 2048
   */
  cursor?: string;
  /**
   * @minimum 1
   * @maximum 200
   */
  limit?: number;
  from?: string;
  to?: string;
  caseId?: string;
  conversationId?: string;
  operatorCmsUserId?: string;
};